refactor: ajustar a locaÃlização dos componentes (sistema de arquivos)
